Add additional information to the Command not found Error.
Changes to be committed:
	modified:   nbqa/__main__.py
	modified:   nbqa/cmdline.py
	deleted:    nbqa/troubleshoot.py
	modified:   tests/test_cmdline.py
	modified:   tests/test_runtime_errors.py
	deleted:    tests/test_troubleshoot.py
